requirements.txt
Flask
requests
beautifulsoup4
python-dotenv
python-telegram-bot~=21.0